feat: implement eventbus publisher/subscriber
