--- backend/app/api_gateway/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class ApiGatewayConfig(AppConfig):
-    name = 'api_gateway'
