--- backend/app/common/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class CommonConfig(AppConfig):
-    name = 'common'
