--- backend/app/common/notification_publisher.py
@@ -1,7 +1,0 @@
-from interface import Interface
-
-
-class NotificationPublisher(Interface):
-
-    def publish_notification(self):
-        pass
