+++ backend/common/command/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class CommandConfig(AppConfig):
+    name = 'command'
