+++ backend/common/messaging/message_handler.py
@@ -1,0 +1,7 @@
+from interface import Interface
+
+
+class MessageHandler(Interface):
+
+    def handle(self, message):
+        pass
