+++ backend/common/messaging/message_publisher.py
@@ -1,0 +1,7 @@
+from interface import Interface
+
+
+class MessagePublisher(Interface):
+
+    def publish_message(self, message):
+        pass
