+++ backend/common/messaging/message_subscriber.py
@@ -1,0 +1,10 @@
+from interface import Interface
+
+
+class MessageSubscriber(Interface):
+
+    def subscribe_message(self, topic, message_handler):
+        pass
+
+    def close(self):
+        pass
