+++ backend/common/utils/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class UtilsConfig(AppConfig):
+    name = 'utils'
