+++ backend/requirements/development.txt
@@ -5,6 +5,7 @@ python-interface
 redis>=3.3.11,<3.3.12
 django-environ
 tenacity>=5.1.1,<5.1.2
+pytest>=5.2.2,<5.2.3
 
 flake8>=3.6.0,<3.7.0
 autopep8>1.4.0,<=1.4.4
